Integrate multiple SRT data

Integrate multiple SRT data based on the PRECASTObj by PRECAST model fitting.

IntegrateSpaData(PRECASTObj, species="Human", 
                 custom_housekeep=NULL, covariates_use=NULL)

Arguments

PRECASTObj

a PRECASTObj object after finishing the PRECAST model fitting and model selection.

species

an optional string, one of 'Human', 'Mouse' and 'Unknown', specify the species of the SRT data to help choose the housekeeping genes. 'Unknown' means only using the PRECAST results reconstruct the alligned gene expression.

custom_housekeep

user-specified housekeeping genes.

covariates_use

a string vector, the colnames in `PRECASTObj@seulist[[1]]@meta.data`, representing other biological covariates to considered when removing batch effects. This is achieved by adding additional covariates for biological conditions in the regression, such as case or control. Default as `NULL`, denoting no other covariates to be considered.

Details

Nothing

Value

Return a Seurat object by integrating all SRT data batches into a SRT data, where the column "batch" in the meta.data represents the batch ID, and the column "cluster" represents the clusters obtained by PRECAST.

References

Wei Liu, Liao, X., Luo, Z. et al, Jin Liu* (2023). Probabilistic embedding, clustering, and alignment for integrating spatial transcriptomics data with PRECAST. Nature Communications, 14, 296

Gagnon-Bartsch, J. A., Jacob, L., & Speed, T. P. (2013). Removing unwanted variation from high dimensional data with negative controls. Berkeley: Tech Reports from Dep Stat Univ California, 1-112.

Author

Wei Liu

Note

nothing

See also

Examples

  data(PRECASTObj)
  PRECASTObj <- selectModel.PRECASTObj(PRECASTObj)
#> Error in selectModel.PRECASTObj(PRECASTObj): could not find function "selectModel.PRECASTObj"
  seuInt <- IntegrateSpaData(PRECASTObj, species='unknown')
#> Using only PRECAST results to obtain the batch corrected gene expressions since species is unknown or the genelist in PRECASTObj has less than 5 overlapp with the housekeeping genes of given species.
#> Users can specify the custom_housekeep by themselves to use the housekeeping genes based methods.
#> Error in matrix(1, nrow = nrow(hV0), ncol = 1): non-numeric matrix extent