Spatial heatmap

Plot spatial heatmap for a Seurat object with spatial transcriptomics data.

featurePlot(seu, feature=NULL, cols=NULL, pt_size=1, title_size =16, quant=0.5, 
  assay='RNA' , reduction="position")

Arguments

seu

an object named "Seurat". The object of class "Seurat" must include slot "scale.data".

feature

an optional positive integer, specify the number of features to be extracted.

cols

colors used in the plot

pt_size

the size of point in the spatial heatmap plot.

title_size

the title size used for the plot.

quant

the quantile value to generate the gradient color map.

assay

the assay selected for plot.

reduction

the Reduc object for plot.

Details

Nothing

Value

For Seurat object, return a Seurat object. For objcet "matrix" and "dgCMatrix", return a object "matrix" with rownames same as the colnames of X, and colnames "WPCA1" to "WPCAq".

Author

Wei Liu

Note

nothing

See also

Examples

if (FALSE) {
  data(PRECASTObj)
  PRECASTObj <- selectModel(PRECASTObj)
  seuInt <- IntegrateSpaData(PRECASTObj, species='unknown')
  seuInt <- ScaleData(seuInt)
  featurePlot(seuInt, assay='PRE_CAST')
  }