Generate simulated data

Generate simulated spatial transcriptomics data with two data batches.

gendata_seulist(height1=30, width1=30,height2=height1, width2=width1, 
                           p =100, q=10, K=7,  G=4, beta=1.2, sigma2=1, 
                           alpha=8, seed=1, view=FALSE)

Arguments

height1, width1, height2, width2

Height and width of lattice grids for generating spatial coordinates, n=height * width spots for expression matrix data

p

number of genes to generate.

q

number of true latent features to generate gene expression

K

number of clusters (spatial domains/cell types).

G

the number of neighbors. The latter must be one of G = 4 or G = 8, which respectively correspond to a first order and a second order dependency structure. By default, G = 4.

beta

the smoothing parameter in Potts model.

sigma2

Variance of error term in probabilitic PCA model.

alpha

a positive factor of mixture mean values.

seed

random seed for generate data

view

Logical value indicating whether the draw should be printed. Do not display the optional borders.

Details

Nothing

Value

return a list consisting of Seurat objects, where each object is a SRT data batch, and the metadata of this Seurat object will include two columns with names "row" and "col" which are the spatial coordinates.

References

None

Author

Wei Liu

Note

nothing

See also

Examples


  ## we generate two spatial transcriptomics data 
  seulist <- gendata_seulist(height1=20, width1=20,p=200, K=4)
  seulist
#> [[1]]
#> An object of class Seurat 
#> 200 features across 400 samples within 1 assay 
#> Active assay: RNA (200 features, 0 variable features)
#>  1 dimensional reduction calculated: trueEmbed
#> 
#> [[2]]
#> An object of class Seurat 
#> 200 features across 400 samples within 1 assay 
#> Active assay: RNA (200 features, 0 variable features)
#>  1 dimensional reduction calculated: trueEmbed
#>