PRECAST model setting

Set the PRECAST model structure and paramters in the algorithm.

model_set(Sigma_equal=FALSE, Sigma_diag=TRUE,mix_prop_heter=TRUE,
                      error_heter=TRUE, Sp2=TRUE, wpca_int=FALSE,int.model='EEE',
                      coreNum = 1, coreNum_int=coreNum,
                      beta_grid=seq(0.2,4, by=0.2),
                      maxIter_ICM=6,maxIter=20, epsLogLik=1e-5, verbose=TRUE, seed=1)

Arguments

Sigma_equal

an optional logical value, specify whether Sigmaks are equal, default as FALSE.

Sigma_diag

an optional logical value, specify whether Sigmaks are diagonal matrices, default as TRUE.

mix_prop_heter

an optional logical value, specify whether betar are distict, default as TRUE.

error_heter

an optional logical value, whether use the heterogenous error i.e. lambdarj != lambdark for each sample r, default as TRUE. If error_heter=FALSE, then the homogenuous error is used for probabilistic PCA model.

Sp2

an optional logical value, whether add the ICAR model component in the model, default as TRUE. We provide this interface for those users who don't want to include the ICAR model.

wpca.int

an optional logical value, means whether use the weighted PCA to obtain the initial values of loadings and other paramters, default as FALSE which means the ordinary PCA is used.

int.model

an optional string, specify which Gaussian mixture model is used in evaluting the initial values for PRECAST, default as "EEE"; and see Mclust for more models' names.

coreNum

an optional positive integer, means the number of thread used in parallel computating.

coreNum_int

an optional positive integer, means the number of cores used in parallel computation for initial values when K is a vector, default as same as coreNum.

beta_grid

an optional vector of positive value, the candidate set of the smoothing parameter to be searched by the grid-search optimization approach.

maxIter_ICM

an optional positive value, represents the maximum iterations of ICM.

maxIter

an optional positive value, represents the maximum iterations of EM.

epsLogLik

an optional positive vlaue, tolerance vlaue of relative variation rate of the observed pseudo log-loglikelihood value, defualt as '1e-5'.

verbose

an optional logical value, whether output the information of the ICM-EM algorithm.

seed

an optional integer, the random seed in fitting PRECAST model.

Details

Nothing

Value

Return a list including all paramters' setting.

Author

Wei Liu

Note

nothing

See also

Examples

  model_set()
#> $Sigma_equal
#> [1] FALSE
#> 
#> $Sigma_diag
#> [1] TRUE
#> 
#> $mix_prop_heter
#> [1] TRUE
#> 
#> $error_heter
#> [1] TRUE
#> 
#> $Sp2
#> [1] TRUE
#> 
#> $wpca_int
#> [1] FALSE
#> 
#> $int.model
#> [1] "EEE"
#> 
#> $coreNum
#> [1] 1
#> 
#> $coreNum_int
#> [1] 1
#> 
#> $beta_grid
#>  [1] 0.2 0.4 0.6 0.8 1.0 1.2 1.4 1.6 1.8 2.0 2.2 2.4 2.6 2.8 3.0 3.2 3.4 3.6 3.8
#> [20] 4.0
#> 
#> $maxIter_ICM
#> [1] 6
#> 
#> $maxIter
#> [1] 20
#> 
#> $epsLogLik
#> [1] 1e-05
#> 
#> $verbose
#> [1] TRUE
#> 
#> $seed
#> [1] 1
#>