Spatial heatmap

Plot spatial heatmap for a Seurat object with spatial transcriptomics data.

ridgePlot(object, features, group.names='Domain' , cols = NULL, idents = NULL, sort = FALSE, 
                         assay = NULL, group.by = NULL, y.max = NULL, same.y.lims = FALSE, 
                         log = FALSE, ncol = NULL, slot = "data", stack = FALSE, 
                         combine = TRUE, fill.by = "feature")

Arguments

object

an object named "Seurat".

features

an optional positive integer, specify the number of features to be extracted.

group.names

the size of point in the scatter plot.

cols

colors to use for plotting.

idents

which classes to include in the plot (default is all)

sort

sort identity classes (on the x-axis) by the average expression of the attribute being potted, can also pass 'increasing' or 'decreasing' to change sort direction.

assay

name of assay to use, defaults to the active assay.

group.by

group (color) cells in different ways.

y.max

maximum y axis value

same.y.lims

set all the y-axis limits to the same values

log

plot the feature axis on log scale

ncol

number of columns if multiple plots are displayed

slot

use non-normalized counts data for plotting

stack

horizontally stack plots for each feature

combine

combine plots into a single patchworked ggplot object. If FALSE, return a list of ggplot

fill.by

color violins/ridges based on either 'feature' or 'ident'

Details

Nothing

Value

Return a single patchworked ggplot object. If coombine = FALSE, return a list of ggplot.

Author

Wei Liu

Note

nothing

See also

Examples

if (FALSE) {
  data(PRECASTObj)
  PRECASTObj <- selectModel(PRECASTObj)
  seuInt <- IntegrateSpaData(PRECASTObj, species='unknown')
  ridgePlot(seuInt, features=row.names(seuInt)[1:4], ncol=2)
  }