ICM-EM algorithm implementation with organized paramters

Efficient data integration as well as spatial clustering for multiple spatial transcriptomics data

ICM.EM_structure(XList,  K, AdjList, q=15,parameterList=NULL)

Arguments

XList

an M-length list consisting of multiple matrices with class dgCMatrix or matrix that specify the log-normalization gene expression matrix for each data sample used for PRECAST model.

K

a positive integer allowing scalar or vector, specify the number of clusters in model fitting.

AdjList

an M-length list of sparse matrices with class dgCMatrix, specify the adjacency matrix used for Potts model and Intrisic CAR model in PRECAST model. We provide this interface for those users who would like to define the adjacency matrix by their own.

q

a positive integer, specify the number of latent features to be extracted, default as 15.

parameterList

Other arguments in PRECAST model, it can be set by model_set.

Details

Nothing

Value

ICM.EM_structure returns a list with class "SeqKiDRSC_Object" with the number of components equal to the length of K, where each component includes the model fitting results for one number of cluster and is a list consisting of following components:

cluster

an M-length list that includes the inferred class labels for each data sample.

hZ

an M-length list that includes the batch corrected low-dimensional embeddings for each data sample.

hV

an M-length list that includes the estimate the ICAR component for each sample.

Rf

an M-length list that includes the posterior probability of domain clusters for each sample.

beta

an M-length vector that includes the estimated smoothing parameters for each sample.

Mu

mean vectors of mixtures components.

Sigma

covariance matrix of mixtures components.

W

estimated loading matrix

Lam

estimated variance of errors in probabilistic PCA model

loglik

pseudo observed log-likelihood.

References

Wei Liu, Xu Liao, Yi Yang, Huazhen Lin, Joe Yeong, Xiang Zhou, Xingjie Shi and Jin Liu. (2022) Joint dimension reduction and clustering analysis for single-cell RNA-seq and spatial transcriptomics data

Author

Wei Liu

Note

nothing

See also

Examples

if (FALSE) {
  ## we generate the spatial transcriptomics data with lattice neighborhood, i.e. ST platform.
  q <- 10; K <- 4
  dat <- PRECAST:::gendataInte_sp(width1 = 20, height1 = 20,height2=21,
  width2=21, p=100,q=q, K=K, sigma2 = 9,beta=c(0.5, 1.0))
  AdjList <- lapply(list(dat$pos1, dat$pos2), getAdj_reg, platform='ST')
  message(sapply(AdjList, dim))
  resList <- ICM.EM_structure(list(dat$X1, dat$X2),  AdjList = AdjList, 
                   q=q, K=K)
}